test(react-compiler): move accessor coverage to fixtures
