+++ crates/oxc_react_compiler/tests/transform.rs
@@ -135,165 +135,6 @@ export function Component({ value }) {
 }
 
 #[test]
-fn object_accessors_preserve_manual_memoization_calls() {
-    let source = "\
-import { useCallback, useMemo } from 'react';
-export function Component({ value }) {
-  const callback = useCallback(() => value, [value]);
-  const memo = useMemo(() => ({ value }), [value]);
-  const object = {
-    get value() {
-      return memo.value;
-    },
-  };
-  return <button onClick={callback}>{object.value}</button>;
-}
-";
-
-    let allocator = Allocator::default();
-    let (program, result) =
-        transform_source(source, SourceType::tsx(), &allocator, PluginOptions::default());
-
-    assert!(result.changed, "component should compile: {:?}", result.diagnostics);
-    assert!(result.diagnostics.is_empty(), "unexpected diagnostics: {:?}", result.diagnostics);
-
-    let output = Codegen::new().build(&program).code;
-    assert!(
-        output.contains("useCallback(") && output.contains("useMemo("),
-        "React must retain manual memoization when compiler caching is disabled:\n{output}"
-    );
-    assert!(
-        !output.contains("react/compiler-runtime"),
-        "accessor-bearing functions must not use compiler caching:\n{output}"
-    );
-}
-
-#[test]
-fn nested_object_accessors_disable_parent_caching() {
-    let source = "\
-import { useMemo } from 'react';
-export function Component({ value }) {
-  const memo = useMemo(() => value, [value]);
-  const read = () => {
-    const object = {
-      get value() {
-        return value;
-      },
-    };
-    return object.value;
-  };
-  return <div>{memo + read()}</div>;
-}
-";
-
-    let allocator = Allocator::default();
-    let (program, result) =
-        transform_source(source, SourceType::tsx(), &allocator, PluginOptions::default());
-
-    assert!(result.changed, "component should compile: {:?}", result.diagnostics);
-    assert!(result.diagnostics.is_empty(), "unexpected diagnostics: {:?}", result.diagnostics);
-
-    let output = Codegen::new().build(&program).code;
-    assert!(output.contains("useMemo("), "manual memoization must be retained:\n{output}");
-    assert!(
-        !output.contains("react/compiler-runtime"),
-        "nested accessors must disable parent compiler caching:\n{output}"
-    );
-}
-
-#[test]
-fn object_accessors_keep_generated_names_unique() {
-    let source = "\
-export function Component({ value }) {
-  const first = { method(y) { return y; } };
-  const x = value;
-  const object = {
-    get value() {
-      return x;
-    },
-  };
-  return <div>{object.value + first.method(1)}</div>;
-}
-";
-
-    let allocator = Allocator::default();
-    let (program, result) =
-        transform_source(source, SourceType::tsx(), &allocator, PluginOptions::default());
-
-    assert!(result.changed, "component should compile: {:?}", result.diagnostics);
-    assert!(result.diagnostics.is_empty(), "unexpected diagnostics: {:?}", result.diagnostics);
-
-    let output = Codegen::new().build(&program).code;
-    let output_allocator = Allocator::default();
-    let parsed = Parser::new(&output_allocator, &output, SourceType::tsx()).parse();
-    assert!(parsed.diagnostics.is_empty(), "compiled accessor output must parse:\n{output}");
-    let semantic = SemanticBuilder::new().build(&parsed.program);
-    assert!(
-        semantic.diagnostics.is_empty(),
-        "compiled accessor output must not contain duplicate declarations:\n{output}"
-    );
-}
-
-#[test]
-fn object_accessors_preserve_typescript_annotations() {
-    let source = "\
-type Model = { value: number };
-export function Component({ value }: { value: number }) {
-  const object = {
-    get value(this: Model): number {
-      return value;
-    },
-    set value(this: Model, next: number) {},
-  };
-  return <div>{object.value}</div>;
-}
-";
-
-    let allocator = Allocator::default();
-    let (program, result) =
-        transform_source(source, SourceType::tsx(), &allocator, PluginOptions::default());
-
-    assert!(result.changed, "component should compile: {:?}", result.diagnostics);
-    assert!(result.diagnostics.is_empty(), "unexpected diagnostics: {:?}", result.diagnostics);
-
-    let output = Codegen::new().build(&program).code;
-    assert!(
-        output.contains("get value(this: Model): number")
-            && output.contains("set value(this: Model, next: number)"),
-        "accessor TypeScript annotations must be preserved:\n{output}"
-    );
-}
-
-#[test]
-fn object_accessors_rename_typescript_annotation_references() {
-    let source = "\
-export function Component({ value }: { value: number }) {
-  const identity = (x: number) => x;
-  const x = value;
-  const object = {
-    get value(): typeof x {
-      return x;
-    },
-  };
-  return <div>{identity(object.value)}</div>;
-}
-";
-
-    let allocator = Allocator::default();
-    let (program, result) =
-        transform_source(source, SourceType::tsx(), &allocator, PluginOptions::default());
-
-    assert!(result.changed, "component should compile: {:?}", result.diagnostics);
-    assert!(result.diagnostics.is_empty(), "unexpected diagnostics: {:?}", result.diagnostics);
-
-    let output = Codegen::new().build(&program).code;
-    assert!(
-        output.contains("get value(): typeof x_0") && output.contains("return x_0"),
-        "accessor annotation references must follow binding renames:\n{output}"
-    );
-}
-
-#[test]
 fn skips_non_react_code() {
     let source = "function add(a, b) {\n  return a + b;\n}\n";
     let allocator = Allocator::default();
